project/requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.3.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.0
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.4
django-appconf==1.2.0
django-baton==5.0.3
django-cachalot==2.8.0
django-filter==23.5
django-imagekit==4.1.0
django-redis==5.4.0
django-rest-passwordreset==1.5.0
django-silk==5.4.3
djangorestframework==3.15.1
drf-spectacular==0.27.1
easy-thumbnails==2.9
gprof2dot==2025.4.14
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
oauthlib==3.3.1
packaging==25.0
pilkit==3.0
pillow==10.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.0.1
referencing==0.37.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.30.0
sentry-sdk==1.45.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.7.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
